Create submodule for constraints code
